Fix BoltProvider not deleting data from tables before inserting new data
This is needed as wirwl only tracks the current state of entries and
their types. When deletion/change to any of them is done, it's only on
the working data. Database changes are only made when user saves but at
this point application doesn't know what was removed/changed so it just
needs to save the working data as it is, without caring what was saved
in db previously.