Add wrapper functions over TabContainer functions used in application
Those fyne's TabContainer functions are used in the application but
because TabContainer cannot be extended right now due to the bug, it
makes sense to wrap the internal container variable 'it' in functions
instead of exposing 'it' directly to access them.
This is because when extending becomes possible,
it will be less costly to update applications code to reflect the fix.

Basically, there will be only need to change Items() to Items instead of
removing .It access everywhere.