Remove workaround for tabs buttons not refreshing in TabContainer
In version 1.2.4 of fyne this was fixed.

A couple of important notes on some changes here:
fyne's TabContainer now can't be a pointer as widgets have to be
extended using variables.

Items need to be placed in TabContainer after ExtendBaseWidget gets
called, otherwise widget doesn't work correctly.

Items function has not been removed as it's useful for encapsulating
the Items variable from fyne's TabContainer.