Modify key press simulation in App tests to all use App's SimulateKeyPress
It was cumbersome to call key press simulation on a correct object that
is currently focused since SimulateKeyPress on App only worked if key
was pressed on the Canvas itself.
But since it's possible to return currently focused object from Canvas,
it is possible to modify this and SimulateKeyPress now passes the key
even to focused object, if such exists.