Simplify FormDialog's ItemValue getter to not return error
Error is never checked as it is assumed that call uses a proper item
name. In case the name is wrong, empty string is returned which in case
that non empty value was expected, can be used to check the correctness
of the returned value.

This also makes calling this function less cumbersome.