Make application load any settings from changeable paths in one place
Previously, there was a path passed in wirlw app's constructor and fyne's app
was pased in on load. Now, to avoid to have loading split between those
two functions, the constructor only takes the app and all actual loading
is in the loading function.
Also in this loading function now all paths that are used to load various
settings (logging and config file) can be passed, which means they are in
one place and can also be replaced for easier testing.