Make application configure itself using AppConfigurator
This allows to prepare and pass all needed dependencies into the App's constructor
which greatly helps in seting up specific situations when testing. For
example, it was hard to test some error handling because App's
dependencies were loaded internally e.g. Provider and it was not
possible to setup a version behaving incorrectly because it would be
always overwritten when App was loading. Having it be loaded outside and
passed in ready to go in a constructor, solves the problem.