Handle input in input field in input mode
This is so typing in input fields behaves more like vim's input mode.
This is for example so that pressing 'jj' (common vim Escape override)
correctly exits the input in every case.
The other input.handler's mode works with either single keys
or combinations but it doesn't allow to always treat two last pressed
characters as a combination which is what is exactly needed for this
common overwrite to work.