Add sufix to entries table name to prevent table name colisions
As entries tables names were simply called after entry type name so it was
technicaly possible for user to create an entry type with name
entries_types which would colide with table for entries types (it is
called like that).
Also, this is usefull in case in the future more tables were to be
added.