Make FormDialog embedded FocusableDialog as pointer
This is needed from now on any widget that extends internal widgets so
that callback functions that might be called by the parent work
properly. If embed is a value then a copy of calling function is made
and because it is not set on construction but later, it can't be
changed.
So for example, if one were to embed FormDialog somewhere, without the
pointer a call to handleEnterKey() from confirm action wouldn't work for
the reason stated above.