Fix table and it's renderer tests not using the correct renderer in assertions
There were two renderers, the one created during table creation and the
one that was created internally by fyne, which makes sense as fyne has
to create a renderer somehow. The problem is the manual one was used for
assertions which is not the same on which the application code operates.
This means some assertions would fail.
The correct way of doing this is to use fyne's test.WidgetRenderer which
will access the actual renderer that was created internally.